+++ src/app/review/_components/delete-button/DeleteButton.tsx
@@ -1,0 +1,10 @@
+import DeleteIcon from '@/components/icons/DeleteIcon';
+
+export const DeleteButton = () => {
+  return (
+    <div className="flex items-center h-8 gap-1 self-end">
+      <DeleteIcon size={20} />
+      <p>삭제</p>
+    </div>
+  );
+};
